+++ src/game/server/tf/bot/behavior/tf_bot_get_ammo.cpp
@@ -18,6 +18,8 @@ ConVar tf_bot_ammo_search_range( "tf_bot_ammo_search_range", "5000", FCVAR_CHEAT
 ConVar tf_bot_crumpkin_search_range( "tf_bot_crumpkin_search_range", "600", FCVAR_CHEAT, "How far bots will search to find crumpkins around them" );
 ConVar tf_bot_spell_search_range( "tf_bot_spell_search_range", "1200", FCVAR_CHEAT, "How far bots will search to find spells around them" );
 ConVar tf_bot_powerup_search_range( "tf_bot_powerup_search_range", "1200", FCVAR_CHEAT, "How far bots will search to find powerups around them" );
+ConVar tf_bot_credits_search_range( "tf_bot_credits_search_range", "2400", FCVAR_CHEAT, "How far bots will search to find credits around them" );
+ConVar tf_bot_core_search_range( "tf_bot_core_search_range", "2000", FCVAR_CHEAT, "How far bots will search to find cores around them" );
 ConVar tf_bot_debug_ammo_scavenging( "tf_bot_debug_ammo_scavenging", "0", FCVAR_CHEAT );
 
 
@@ -31,6 +33,9 @@ CTFBotGetAmmo::CTFBotGetAmmo( void )
 	m_isGoalSpell = false;
 	m_isGoalPowerup = false;
 	m_isGoalMerasmus = false;
+	m_isGoalCredits = false;
+	m_isGoalCores = false;
+	m_isGoalGeneric = false;
 }
 
 CTFBotGetAmmo::CTFBotGetAmmo( bool crumpkin )
@@ -39,10 +44,12 @@ CTFBotGetAmmo::CTFBotGetAmmo( bool crumpkin )
 	m_ammo = NULL;
 	m_isGoalDispenser = false;
 	m_isGoalCrumpkin = crumpkin;
-	m_isGoalSpell = !crumpkin;
+	m_isGoalSpell = false;
 	m_isGoalPowerup = false;
 	m_isGoalGeneric = false;
 	m_isGoalMerasmus = false;
+	m_isGoalCredits = false;
+	m_isGoalCores = false;
 }
 
 CTFBotGetAmmo::CTFBotGetAmmo( bool crumpkin, bool powerup )
@@ -52,9 +59,11 @@ CTFBotGetAmmo::CTFBotGetAmmo( bool crumpkin, bool powerup )
 	m_isGoalDispenser = false;
 	m_isGoalCrumpkin = false;
 	m_isGoalSpell = false;
-	m_isGoalPowerup = true;
-	m_isGoalGeneric = false;
+	m_isGoalPowerup = false;
+	m_isGoalGeneric = true;
 	m_isGoalMerasmus = false;
+	m_isGoalCredits = false;
+	m_isGoalCores = false;
 }
 
 //---------------------------------------------------------------------------------------------
@@ -68,6 +77,8 @@ public:
 		m_crumpkin = false;
 		m_spell = false;
 		m_powerup = false;
+		m_credits = false;
+		m_cores = false;
 	}
 
 	bool IsSelected( const CBaseEntity *constCandidate ) const
@@ -88,6 +99,16 @@ public:
 		if ( m_spell )
 		{
 			return candidate->ClassMatches( "tf_spell_pickup" ) && !candidate->IsEffectActive( EF_NODRAW );
+		}
+
+		if ( m_credits )
+		{
+			return candidate->ClassMatches( "item_currencypack*" ) && !candidate->IsEffectActive( EF_NODRAW );
+		}
+
+		if ( m_cores )
+		{
+			return candidate->ClassMatches( "item_bonuspack" ) && !candidate->IsEffectActive( EF_NODRAW ) && candidate->GetTeamNumber() == m_me->GetTeamNumber();
 		}
 
 		if ( m_powerup )
@@ -169,6 +190,8 @@ public:
 	bool m_crumpkin;
 	bool m_spell;
 	bool m_powerup;
+	bool m_credits;
+	bool m_cores;
 };
 
 
@@ -188,6 +211,8 @@ CTFBotGetAmmo::CTFBotGetAmmo( CTFBot* me, CBaseEntity* target )
 	m_isGoalPowerup = false;
 	m_isGoalGeneric = true;
 	m_isGoalMerasmus = false;
+	m_isGoalCredits = false;
+	m_isGoalCores = false;
 	s_possibleBot = me;
 	s_possibleAmmo = target;
 	s_possibleFrame = gpGlobals->framecount;
@@ -432,6 +457,109 @@ bool CTFBotGetAmmo::IsPowerupPossible( CTFBot* me )
 	return true;
 }
 
+bool CTFBotGetAmmo::IsCreditPossible( CTFBot* me )
+{
+	VPROF_BUDGET( "CTFBotGetAmmo::IsCreditPossible", "NextBot" );
+
+	int i;
+
+	CUtlVector< CNavArea* > nearbyAreaVector;
+	CollectSurroundingAreas( &nearbyAreaVector, me->GetLastKnownArea(), tf_bot_credits_search_range.GetFloat(), me->GetLocomotionInterface()->GetStepHeight(), me->GetLocomotionInterface()->GetDeathDropHeight() );
+
+	CAmmoFilter ammoFilter( me );
+	ammoFilter.m_credits = true;
+	CBaseEntity* closestAmmo = NULL;
+	float closestAmmoTravelDistance = FLT_MAX;
+
+	// append nearby crumpkins
+	CBaseEntity* ammoPack = NULL;
+	while ( ( ammoPack = gEntList.FindEntityByClassname( ammoPack, "item_currencypack*" ) ) != NULL )
+	{
+		if ( ammoFilter.IsSelected( ammoPack ) )
+		{
+			if ( ammoFilter.m_ammoArea && ammoFilter.m_ammoArea->IsMarked() )
+			{
+				if ( ammoFilter.m_ammoArea->GetCostSoFar() < closestAmmoTravelDistance )
+				{
+					closestAmmo = ammoPack;
+					closestAmmoTravelDistance = ammoFilter.m_ammoArea->GetCostSoFar();
+				}
+
+				if ( tf_bot_debug_ammo_scavenging.GetBool() )
+				{
+					NDebugOverlay::Cross3D(ammoPack->WorldSpaceCenter(), 5.0f, 255, 100, 0, true, 999.9);
+				}
+			}
+		}
+	}
+
+	if ( !closestAmmo )
+	{
+		//if ( me->IsDebugging( NEXTBOT_BEHAVIOR ) )
+		//{
+		//	Log( "%3.2f: No powerup nearby\n", gpGlobals->curtime );
+		//}
+		return false;
+	}
+
+	s_possibleBot = me;
+	s_possibleAmmo = closestAmmo;
+	s_possibleFrame = gpGlobals->framecount;
+
+	return true;
+}
+
+bool CTFBotGetAmmo::IsCorePossible( CTFBot* me )
+{
+	VPROF_BUDGET( "CTFBotGetAmmo::IsCorePossible", "NextBot" );
+
+	int i;
+
+	CUtlVector< CNavArea* > nearbyAreaVector;
+	CollectSurroundingAreas( &nearbyAreaVector, me->GetLastKnownArea(), tf_bot_core_search_range.GetFloat(), me->GetLocomotionInterface()->GetStepHeight(), me->GetLocomotionInterface()->GetDeathDropHeight() );
+
+	CAmmoFilter ammoFilter( me );
+	ammoFilter.m_cores = true;
+	CBaseEntity* closestAmmo = NULL;
+	float closestAmmoTravelDistance = FLT_MAX;
+
+	// append nearby crumpkins
+	CBaseEntity* ammoPack = NULL;
+	while ( ( ammoPack = gEntList.FindEntityByClassname( ammoPack, "item_bonuspack" ) ) != NULL )
+	{
+		if ( ammoFilter.IsSelected( ammoPack ) )
+		{
+			if ( ammoFilter.m_ammoArea && ammoFilter.m_ammoArea->IsMarked() )
+			{
+				if ( ammoFilter.m_ammoArea->GetCostSoFar() < closestAmmoTravelDistance )
+				{
+					closestAmmo = ammoPack;
+					closestAmmoTravelDistance = ammoFilter.m_ammoArea->GetCostSoFar();
+				}
+
+				if ( tf_bot_debug_ammo_scavenging.GetBool() )
+				{
+					NDebugOverlay::Cross3D(ammoPack->WorldSpaceCenter(), 5.0f, 255, 100, 0, true, 999.9);
+				}
+			}
+		}
+	}
+
+	if ( !closestAmmo )
+	{
+		//if ( me->IsDebugging( NEXTBOT_BEHAVIOR ) )
+		//{
+		//	Log( "%3.2f: No powerup nearby\n", gpGlobals->curtime );
+		//}
+		return false;
+	}
+
+	s_possibleBot = me;
+	s_possibleAmmo = closestAmmo;
+	s_possibleFrame = gpGlobals->framecount;
+
+	return true;
+}
 
 //---------------------------------------------------------------------------------------------
 ActionResult< CTFBot >	CTFBotGetAmmo::OnStart( CTFBot *me, Action< CTFBot > *priorAction )
@@ -452,6 +580,10 @@ ActionResult< CTFBot >	CTFBotGetAmmo::OnStart( CTFBot *me, Action< CTFBot > *pri
 	m_ammo = s_possibleAmmo;
 	m_isGoalDispenser = m_ammo->ClassMatches( "obj_dispenser*" );
 	m_isGoalMerasmus = m_ammo->ClassMatches( "merasmus" );
+	m_isGoalSpell = m_ammo->ClassMatches( "tf_spell_pickup" );
+	m_isGoalPowerup = m_ammo->ClassMatches( "item_powerup_rune" );
+	m_isGoalCredits = m_ammo->ClassMatches( "item_currencypack*" );
+	m_isGoalCores = m_ammo->ClassMatches( "item_bonuspack" );
 
 	CTFBotPathCost cost( me, FASTEST_ROUTE );
 	if ( !m_path.Compute( me, m_ammo->WorldSpaceCenter(), cost ) )
@@ -460,7 +592,7 @@ ActionResult< CTFBot >	CTFBotGetAmmo::OnStart( CTFBot *me, Action< CTFBot > *pri
 	}
 
 	// if I'm a spy, cloak and disguise
-	if ( me->IsPlayerClass( TF_CLASS_SPY ) && !m_isGoalCrumpkin && !m_isGoalPowerup && !m_isGoalSpell && !m_isGoalGeneric )
+	if ( me->IsPlayerClass( TF_CLASS_SPY ) && !m_isGoalCrumpkin && !m_isGoalGeneric )
 	{
 		if ( !me->m_Shared.IsStealthed() )
 		{
@@ -475,7 +607,7 @@ ActionResult< CTFBot >	CTFBotGetAmmo::OnStart( CTFBot *me, Action< CTFBot > *pri
 //---------------------------------------------------------------------------------------------
 ActionResult< CTFBot >	CTFBotGetAmmo::Update( CTFBot *me, float interval )
 {
-	if ( me->IsAmmoFull() && !m_isGoalCrumpkin && !m_isGoalSpell && !m_isGoalPowerup && !m_isGoalGeneric )
+	if ( me->IsAmmoFull() && !m_isGoalCrumpkin && !m_isGoalGeneric )
 	{
 		return Done( "My ammo is full" );
 	}
@@ -491,7 +623,7 @@ ActionResult< CTFBot >	CTFBotGetAmmo::Update( CTFBot *me, float interval )
 	{
 		if ( me->m_Shared.IsCarryingRune() )
 		{
-			return Done("Don't need to look for powerups anymore");
+			return Done( "Don't need to look for powerups anymore" );
 		}
 	}