+++ src/game/server/tf/bot/behavior/tf_bot_tactical_monitor.h
@@ -43,6 +43,8 @@ private:
 	bool ShouldOpportunisticallyCollectCrumpkins( CTFBot* me ) const;
 	bool ShouldOpportunisticallyCollectSpell( CTFBot* me ) const;
 	bool ShouldOpportunisticallyCollectPowerup( CTFBot* me ) const;
+	bool ShouldOpportunisticallyCollectCredits( CTFBot* me ) const;
+	bool ShouldOpportunisticallyCollectCores( CTFBot* me ) const;
 
 	void AvoidBumpingEnemies( CTFBot *me );
 };